Sign In Controller Implementation
ADD - resign the first responder for both email and password.
ADD - retrieve the text for both form values (email and password).
ADD - sign in action method to allow sign in with form values entered by the suer.
ADD - forgot password method to allow the user to reset password with the entered email.
ADD - skip action method to transition directly into the main interface of the application.
ADD - unwind segue to allow the user to sign out of their account and transition back into the sign in screen.