Article Model
ADD - getter/property to format the date into MM, dd, yyyy format.
ADD - init to constructor an article with the expected values.
ADD - convenience init as a helper to assist on parsing articles.
ADD - method to asynchronously cache and download an image from a headline article (not fully tested)
ADD - method to parse the data with date formatter.