AsynchRequest & Categorized Sources
ADD - the async request class which will assist on performing network requests with result and storing that result as needed.
ADD - the categorized sources class, which will assist with categorizing sources if needed through details and categories. In addition, it assist with archiving/unarchiving and encoding/decoding categorized sources objects as needed.