Download Article Refactor & Download Headlines/Articles Implementation
ADD - refactor of the download articles into a dynamic/unique method to be used to download articles and headlines.
ADD - method to download articles based on the given source.
ADD - method to download headlines based on the given country.
ADD - methods to get the result after the download of headlines and articles is finished returns the async request result.