Article Detail View, Article View, Sources View
ADD - article detail view controller to displayed the expected information of an article or headline to the user.
ADD - article view controller to displayed all the possible articles from a chosen source.
ADD -  prepare for segue to the sources view controller to allow the user to transition into all the possible articles for the chosen source.