Article Watch Interface Controllers
ADD - article row controller which will retrieve and set all articles text to be displayed on the screen.
ADD - article interface controller which contains functionality to populate the table of articles and presenting the articles detail view controller.
ADD - article detail interface controller which consists of presenting the article contents to the user.